fix Hex2P 'createPiece' and 'applyPerspective' | also deduplicate createPiece
